+++ src/Livewire/TerminalBuilder.php
@@ -130,6 +130,7 @@ class TerminalBuilder
             'squareCorners' => $this->getSquareCorners(),
             'scripts' => $this->getScripts(),
             'autoConnect' => $this->getAutoConnect(),
+            'connectionBehavior' => $this->getEffectiveConnectionBehavior()->value,
             'loggingEnabled' => $this->getLoggingEnabled(),
             'logConnections' => $this->getLogConnections(),
             'logIdentifier' => $this->getLogIdentifier(),